+++ packages/experimental/src/components/PageHeader/PageHeader-test.js
@@ -7,12 +7,8 @@
 
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
 
-import { expPrefix } from '../../global/js/settings';
-
-// import { settings } from 'carbon-components';
-// const { prefix: carbonPrefix } = settings;
+import { pkgPrefix /*, carbonPrefix */ } from '../../global/js/settings';
 
 import {
   BreadcrumbItem,
@@ -23,7 +19,7 @@ import {
 } from 'carbon-components-react';
 import { Lightning16, Bee32 } from '@carbon/icons-react';
 
-import { ActionBarItem } from './ActionBarItem';
+import { ActionBarItem } from '../ActionBar';
 import { PageHeader } from '.';
 
 /* Test properties. */
@@ -98,21 +94,16 @@ describe('PageHeader', () => {
   test('renders an empty header when no props are set', () => {
     render(<PageHeader />);
 
-    const header = document.querySelector(`section.${expPrefix}-page-header`);
+    const header = document.querySelector(`section.${pkgPrefix}-page-header`);
 
     expect(header).not.toBeNull();
     expect(
-      header.classList.contains(`${expPrefix}-page-header--background`)
+      header.classList.contains(`${pkgPrefix}-page-header--background`)
     ).toBe(false);
     expect(header.classList.contains(classNames[0])).toBe(false);
     expect(header.classList.contains(classNames[1])).toBe(false);
     expect(
-      document.querySelectorAll(`.${expPrefix}-page-header--action-bar`)
-    ).toHaveLength(0);
-    expect(
-      document.querySelectorAll(
-        `[label="Action 1"].${expPrefix}-action-bar-item`
-      )
+      document.querySelectorAll(`.${pkgPrefix}-page-header--action-bar`)
     ).toHaveLength(0);
     expect(
       document.querySelectorAll('span.page-header-test--available-space')
@@ -120,19 +111,19 @@ describe('PageHeader', () => {
     expect(screen.queryAllByTestId('breadcrumbitem')).toHaveLength(0);
     expect(screen.queryAllByTestId('tabs')).toHaveLength(0);
     expect(
-      document.querySelectorAll(`.${expPrefix}-page-header--page-actions`)
+      document.querySelectorAll(`.${pkgPrefix}-page-header--page-actions`)
     ).toHaveLength(0);
     expect(
-      document.querySelectorAll(`.${expPrefix}-page-header--subtitle`)
+      document.querySelectorAll(`.${pkgPrefix}-page-header--subtitle`)
     ).toHaveLength(0);
     expect(screen.queryByText(subtitle)).toBeNull();
     expect(screen.queryAllByTestId('tags')).toHaveLength(0);
     expect(
-      document.querySelectorAll(`.${expPrefix}-page-header--title`)
+      document.querySelectorAll(`.${pkgPrefix}-page-header--title`)
     ).toHaveLength(0);
     expect(screen.queryByText(title)).toBeNull();
     expect(
-      document.querySelectorAll(`.${expPrefix}-page-header--title-icon`)
+      document.querySelectorAll(`.${pkgPrefix}-page-header--title-icon`)
     ).toHaveLength(0);
   });
 
@@ -153,55 +144,44 @@ describe('PageHeader', () => {
       />
     );
 
-    const header = document.querySelector(`section.${expPrefix}-page-header`);
+    const header = document.querySelector(`section.${pkgPrefix}-page-header`);
+    console.log('Look here', header.innerHTML);
 
     expect(header).not.toBeNull();
     expect(
-      header.classList.contains(`${expPrefix}-page-header--background`)
+      header.classList.contains(`${pkgPrefix}-page-header--background`)
     ).toBe(true);
     expect(header.classList.contains(classNames[0])).toBe(true);
     expect(header.classList.contains(classNames[1])).toBe(true);
     expect(
-      document.querySelectorAll(`.${expPrefix}-page-header--action-bar`)
+      document.querySelectorAll(`.${pkgPrefix}-page-header--action-bar`)
     ).toHaveLength(1);
-    expect(
-      document.querySelectorAll(
-        `[label="Action 1"].${expPrefix}-action-bar-item`
-      )
-    ).toHaveLength(1);
-    userEvent.click(
-      document.querySelector(`[label="Action 1"].${expPrefix}-action-bar-item`)
-    );
-    expect(actionBarItemOnClick).toHaveBeenCalledTimes(1);
     expect(
       document.querySelectorAll('span.page-header-test--available-space')
     ).toHaveLength(1);
     expect(
-      document.querySelectorAll(`.${expPrefix}-page-header--breadcrumb`)
+      document.querySelectorAll(`.${pkgPrefix}-page-header--breadcrumb`)
     ).toHaveLength(1);
     expect(screen.getAllByText(/Breadcrumb [1-3]/)).toHaveLength(3);
     expect(screen.queryAllByTestId('tabs')).toHaveLength(1);
     expect(screen.getAllByText(/Tab [1-4]/)).toHaveLength(4);
     expect(
-      document.querySelectorAll(`.${expPrefix}-page-header--page-actions`)
+      document.querySelectorAll(`.${pkgPrefix}-page-header--page-actions`)
     ).toHaveLength(2);
-    expect(screen.getAllByText('Primary button')).toHaveLength(2);
-    userEvent.click(screen.getAllByText('Primary button')[0]);
-    expect(pageActionItemOnClick).toHaveBeenCalledTimes(1);
     expect(
-      document.querySelectorAll(`.${expPrefix}-page-header--subtitle`)
+      document.querySelectorAll(`.${pkgPrefix}-page-header--subtitle`)
     ).toHaveLength(1);
     expect(screen.getByText(subtitle).textContent).toEqual(subtitle);
     expect(screen.queryAllByTestId('tabs')).toHaveLength(1);
     expect(screen.getAllByText('A tag')).toHaveLength(4);
     expect(
-      document.querySelectorAll(`.${expPrefix}-page-header--title`)
+      document.querySelectorAll(`.${pkgPrefix}-page-header--title`)
     ).toHaveLength(1);
     expect(
-      document.querySelector(`.${expPrefix}-page-header--title`).textContent
+      document.querySelector(`.${pkgPrefix}-page-header--title`).textContent
     ).toEqual(title);
     expect(
-      document.querySelectorAll(`.${expPrefix}-page-header--title-icon`)
+      document.querySelectorAll(`.${pkgPrefix}-page-header--title-icon`)
     ).toHaveLength(1);
   });
 });