+++ packages/experimental/src/components/PageHeader/index.js
@@ -6,5 +6,4 @@
 //
 
 export { PageHeader } from './PageHeader';
-export { ActionBarItem } from './ActionBarItem';
 export { PageActionItem } from './PageActionItem';