+++ packages/security/src/ComboButton/ComboButton.stories.js
@@ -1,10 +1,11 @@
 /**
- * Copyright IBM Corp. 2020, 2020
+ * Copyright IBM Corp. 2020, 2021
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
 
+import CloudApp16 from '@carbon/icons-react/lib/cloud-app/16';
 import React from 'react';
 
 import { sectionTitle } from '../../config';
@@ -24,7 +25,7 @@ export default {
 export const Default = () => (
   <ComboButton>
     <ComboButtonItem>ComboButtonItem 1</ComboButtonItem>
-    <ComboButtonItem>ComboButtonItem 2</ComboButtonItem>
+    <ComboButtonItem renderIcon={CloudApp16}>ComboButtonItem 2</ComboButtonItem>
     <ComboButtonItem>ComboButtonItem 3</ComboButtonItem>
   </ComboButton>
 );